+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockFlowerPotBehavior.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockEntityHolderComponent;
+import org.allaymc.api.blockentity.interfaces.BlockEntityFlowerPot;
 
-public interface BlockFlowerPotBehavior extends BlockBehavior {
+public interface BlockFlowerPotBehavior extends BlockBehavior, BlockEntityHolderComponent<BlockEntityFlowerPot> {
 }